complex film list, complex film by id or category

+++ cinema_spring_boot/src/main/java/com/epul/cinema/cinema_spring_boot/repositories/ActorEntityRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ActorEntityRepository extends JpaRepository<ActorEntity, Integer> {
+public interface ActorEntityRepository extends JpaRepository<ActorEntity, Short> {
 }
